@@ -1,10 +1,11 @@
-﻿using System;
+using System;
 using Camunda.Worker;
 using Camunda.Worker.Extensions;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using SampleCamundaWorker.Handlers;
 
@@ -19,7 +20,6 @@ namespace SampleCamundaWorker
 
         public IConfiguration Configuration { get; }
 
-        // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddCamundaWorker(options =>
@@ -40,15 +40,18 @@ namespace SampleCamundaWorker
                         logger.LogInformation("Finished processing of task {Id}", context.Task.Id);
                     });
                 });
+
+            services.AddHealthChecks();
         }
 
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
+
+            app.UseHealthChecks("/health");
         }
     }
 }
